lib multi bump var to size_t

+++ lib/multi.c
@@ -382,7 +382,7 @@ static void multi_addmsg(struct Curl_multi *multi, struct Curl_message *msg)
 }
 
 struct Curl_multi *Curl_multi_handle(size_t hashsize, /* socket hash */
-                                     int chashsize, /* connection hash */
+                                     size_t chashsize, /* connection hash */
                                      size_t dnssize) /* dns hash */
 {
   struct Curl_multi *multi = calloc(1, sizeof(struct Curl_multi));
